Create server handler include findNetworkNamespacePath and connectBridge
